p4i integration Integrate 125146 from 3 to main: Make Sun CC in compat mode be able to compile this file.
Task #38641

Reviewed by Pascal and Volker

[git-p4: depot-paths = "//depot/qt/main/": change = 125147]

+++ src/tools/qdir.cpp
@@ -1169,7 +1169,7 @@ QString QDir::cleanDirPath( const QString &filePath )
 	while ( upLevel-- )
 	    newPath.insert( 0, QString::fromLatin1("/..") );
 	if ( !newPath.isEmpty() )
-	    newPath.remove( (uint)0, 1 );
+	    newPath.remove( (uint)0, (uint)1 );
 	else
 	    newPath = QString::fromLatin1(".");
     } else {
